add new form component
