+++ src/components/Timer/index.js
@@ -1,12 +1,14 @@
 import React from "react";
 import { Typography, Grid } from '@material-ui/core';
 
-export default ({ message }) => {
+import styles from '../../assets/styles';
 
+export default ({ message }) => {
+    const classes = styles();
     return (
         <>
-            <Grid>
-                <Typography variant="h5"> {message}</Typography>
+            <Grid xs={12}>
+                <Typography className={classes.buttonsContainer} variant="h5"> {message}</Typography>
             </Grid>
         </>
     );
